Move several modules to new 'datastore' subpackage.
The base Datastore class is lifted to
`lsst.daf.butler.datastore.Datastore`, but other modules are expected
to be imported directly from their lower-level modules.
