+++ python/lsst/daf/butler/__init__.py
@@ -1,3 +1,30 @@
+# This file is part of daf_butler.
+#
+# Developed for the LSST Data Management System.
+# This product includes software developed by the LSST Project
+# (http://www.lsst.org).
+# See the COPYRIGHT file at the top-level directory of this distribution
+# for details of code ownership.
+#
+# This software is dual licensed under the GNU General Public License and also
+# under a 3-clause BSD license. Recipients may choose which of these licenses
+# to use; please see the files gpl-3.0.txt and/or bsd_license.txt,
+# respectively.  If you choose the GPL option then the following text applies
+# (but note that there is still no warranty even if you opt for BSD instead):
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 """
 Data Access Butler
 """
@@ -21,20 +48,22 @@ from ._deferredDatasetHandle import *
 from ._limited_butler import *
 from ._quantum_backed import *
 from ._topology import *
-from .composites import *
 from .config import *
 from .configSupport import LookupKey
-from .constraints import *
 from .datastore import *
-from .datastoreCacheManager import *
-from .datastoreRecordData import *
+from .datastore.cache_manager import *
+from .datastore.composites import *
+from .datastore.constraints import *
+from .datastore.file_descriptor import *
+from .datastore.file_templates import *
+from .datastore.location import *
+from .datastore.record_data import *
+from .datastore.storage_class_delegate import *
+from .datastore.stored_file_info import *
 from .dimensions import *
 from .exceptions import *
 from .fileDataset import *
-from .fileDescriptor import *
-from .fileTemplates import *
 from .formatter import *
-from .location import *
 
 # Only import ButlerLogRecords
 # ButlerLogRecords is the fundamental type stored in datastores.
@@ -48,8 +77,6 @@ from .quantum import *
 # Import the registry subpackage directly for other symbols.
 from .registry import CollectionSearch, CollectionType, Registry, RegistryConfig
 from .storageClass import *
-from .storageClassDelegate import *
-from .storedFileInfo import *
 from .timespan import *
 from .transfers import RepoExportContext, YamlRepoExportBackend, YamlRepoImportBackend
 from .version import *
