+++ python/lsst/daf/butler/datastore/file_templates.py
@@ -38,16 +38,16 @@ from collections.abc import Iterable, Mapping
 from types import MappingProxyType
 from typing import TYPE_CHECKING, Any
 
-from ._dataset_ref import DatasetRef
-from .config import Config
-from .configSupport import LookupKey, processLookupConfigs
-from .dimensions import DataCoordinate, SkyPixDimension
-from .exceptions import ValidationError
-from .storageClass import StorageClass
+from .._dataset_ref import DatasetRef
+from ..config import Config
+from ..configSupport import LookupKey, processLookupConfigs
+from ..dimensions import DataCoordinate, SkyPixDimension
+from ..exceptions import ValidationError
+from ..storageClass import StorageClass
 
 if TYPE_CHECKING:
-    from ._dataset_type import DatasetType
-    from .dimensions import DimensionUniverse
+    from .._dataset_type import DatasetType
+    from ..dimensions import DimensionUniverse
 
 log = logging.getLogger(__name__)
