+++ python/lsst/daf/butler/datastore/location.py
