+++ python/lsst/daf/butler/datastore/record_data.py
@@ -40,13 +40,13 @@ from lsst.daf.butler._compat import PYDANTIC_V2, _BaseModelCompat
 from lsst.utils import doImportType
 from lsst.utils.introspection import get_full_type_name
 
-from ._dataset_ref import DatasetId
-from .dimensions import DimensionUniverse
-from .persistenceContext import PersistenceContextVars
-from .storedFileInfo import StoredDatastoreItemInfo
+from .._dataset_ref import DatasetId
+from ..dimensions import DimensionUniverse
+from ..persistenceContext import PersistenceContextVars
+from .stored_file_info import StoredDatastoreItemInfo
 
 if TYPE_CHECKING:
-    from .registry import Registry
+    from ..registry import Registry
 
 # Pydantic 2 requires we be explicit about the types that are used in
 # datastore records. Without this UUID can not be handled. Pydantic v1
